extract ScriptSeed- and LessonPlan-specific logic to their own modules
